Moved: tests for `modifier.py` to correct dir
