+++ tests/sparseml/pytorch/optim/test_manager.py
@@ -29,7 +29,7 @@ from tests.sparseml.pytorch.helpers import (
     create_optim_adam,
     create_optim_sgd,
 )
-from tests.sparseml.pytorch.optim.test_modifier import (
+from tests.sparseml.pytorch.sparsification.test_modifier import (
     ModifierTest,
     ScheduledModifierImpl,
 )
