+++ tests/sparseml/pytorch/optim/test_modifier_distillation.py
@@ -24,7 +24,7 @@ from torch.optim import Optimizer
 from sparseml.pytorch.optim import DistillationModifier
 from sparseml.pytorch.sparsification import Modifier, ScheduledModifier
 from tests.sparseml.pytorch.helpers import LinearNet, create_optim_sgd
-from tests.sparseml.pytorch.optim.test_modifier import ScheduledModifierTest
+from tests.sparseml.pytorch.sparsification.test_modifier import ScheduledModifierTest
 
 
 from tests.sparseml.pytorch.helpers import (  # noqa isort:skip
