+++ tests/sparseml/pytorch/optim/test_modifier_epoch.py
@@ -22,7 +22,7 @@ from tests.sparseml.pytorch.helpers import (
     create_optim_adam,
     create_optim_sgd,
 )
-from tests.sparseml.pytorch.optim.test_modifier import ScheduledModifierTest
+from tests.sparseml.pytorch.sparsification.test_modifier import ScheduledModifierTest
 
 
 from tests.sparseml.pytorch.helpers import (  # noqa isort:skip
