+++ tests/sparseml/pytorch/optim/test_modifier_lr.py
@@ -26,7 +26,7 @@ from sparseml.pytorch.optim import (
 )
 from sparseml.pytorch.utils import get_optim_learning_rate
 from tests.sparseml.pytorch.helpers import LinearNet
-from tests.sparseml.pytorch.optim.test_modifier import (
+from tests.sparseml.pytorch.sparsification.test_modifier import (
     ScheduledModifierTest,
     ScheduledUpdateModifierTest,
 )
