+++ tests/sparseml/pytorch/optim/test_modifier_params.py
@@ -28,7 +28,7 @@ from sparseml.pytorch.utils import (
 )
 from sparseml.utils import ALL_TOKEN
 from tests.sparseml.pytorch.helpers import LinearNet, create_optim_sgd
-from tests.sparseml.pytorch.optim.test_modifier import (
+from tests.sparseml.pytorch.sparsification.test_modifier import (
     ScheduledModifierTest,
     ScheduledUpdateModifierTest,
 )
