+++ tests/sparseml/pytorch/sparsification/pruning/test_modifier_pruning_acdc.py
@@ -19,7 +19,7 @@ from torch.optim import SGD
 
 from sparseml.pytorch.sparsification.pruning import ACDCPruningModifier
 from tests.sparseml.pytorch.helpers import LinearNet
-from tests.sparseml.pytorch.optim.test_modifier import (
+from tests.sparseml.pytorch.sparsification.test_modifier import (
     ScheduledModifierTest,
     create_optim_adam,
 )
