+++ tests/sparseml/pytorch/sparsification/pruning/test_modifier_pruning_magnitude.py
@@ -26,7 +26,7 @@ from sparseml.pytorch.sparsification.pruning import (
 from sparseml.pytorch.utils import tensor_sparsity
 from sparseml.utils import FROM_PARAM_TOKEN
 from tests.sparseml.pytorch.helpers import LinearNet
-from tests.sparseml.pytorch.optim.test_modifier import (
+from tests.sparseml.pytorch.sparsification.test_modifier import (
     ScheduledUpdateModifierTest,
     create_optim_adam,
     create_optim_sgd,
