+++ tests/sparseml/pytorch/sparsification/quantization/test_modifier_quantization.py
@@ -19,7 +19,7 @@ from torch.nn import Conv2d, Identity, Linear
 
 from sparseml.pytorch.sparsification import QuantizationModifier
 from tests.sparseml.pytorch.helpers import LinearNet, create_optim_sgd
-from tests.sparseml.pytorch.optim.test_modifier import ScheduledModifierTest
+from tests.sparseml.pytorch.sparsification.test_modifier import ScheduledModifierTest
 
 
 from tests.sparseml.pytorch.helpers import (  # noqa isort:skip
